refactor: Change identifier representation in binary schema format (AIRBNB)
Previously the identifiers section of the binary schema format incorporated the "kind" of an identifier into it (ie, is identifier a type-def or directive-def?).  While compact, if in the future we want to change the representation used by that section (e.g., to use a Trie), this injection of metadata into the table would increase the difficulty of making changes.  So this PR changes the format to have separate sections, one for a "pure" identifier table and the other for the metadata.

(The PR also reorganizes the code slightly, pulling the IO functions into their own package, and it makes some clarifying adjustments to some of the documentation.)

Github-Change-Id: 979620
GitOrigin-RevId: 9f120ce77b42a52316b4034097e8440801354e60